Move PanelHorizontalTransitionContext to its own file
